bugfix: 持久化变量在AST阶段被解析为全局变量

+++ Lyyneheym/Lyyneheym/PlatformCore/VM/PersistContext.cs
@@ -41,7 +41,7 @@ namespace Yuri.PlatformCore.VM
         {
             lock (this)
             {
-                return this.symbols.ContainsKey(varName) ? this.symbols[varName] : 0;
+                return this.symbols.ContainsKey(varName) ? this.symbols[varName] : (double)0;
             }
         }
     }
